// React/app/GerAdv_TS/Acao/Interfaces/interface.Acao.ts
// 100% auto-generated by Source Genesys WebApi Front & Back Creator
// copyright © 2000-2025 Menphis - Sistemas Inteligentes
// This file is part of the Source Genesys project                     
'use client';

export interface IAcao {
// 202501251
    id: number;
 
	justica: number,
	area: number,
	descricao: string,
}

export interface IAcaoFormProps {
  id: number;
  onClose: () => void;
  onSuccess: (registro?: any) => void;
  onError: () => void;
}

export interface IAcaoIncProps {
    id: number;
    onClose: () => void;
    onSuccess: (registro?: any) => void;
    onError: () => void;
}